+++ src/main/java/module-info.java
@@ -1,14 +1,14 @@
-module jfxtras {
+module tjfxtras {
     requires javafx.controls;
     requires javafx.graphics;
     requires kotlin.stdlib;
     requires annotations;
     //requires org.junit.jupiter.engine;
 
-    opens systems.terranatal.jfxtras;
-    opens systems.terranatal.jfxtras.datautils;
+    opens systems.terranatal.tjfxtras;
+    opens systems.terranatal.tjfxtras.datautils;
 
-    exports systems.terranatal.jfxtras.datautils;
+    exports systems.terranatal.tjfxtras.datautils;
 
-    exports systems.terranatal.jfxtras;
+    exports systems.terranatal.tjfxtras;
 }
