+++ src/main/java/systems/terranatal/tjfxtras/PaneInit.java
@@ -1,4 +1,4 @@
-package systems.terranatal.jfxtras;
+package systems.terranatal.tjfxtras;
 
 import javafx.scene.layout.Pane;
