+++ src/main/java/systems/terranatal/tjfxtras/datautils/Converters.java
@@ -1,4 +1,4 @@
-package systems.terranatal.jfxtras.datautils;
+package systems.terranatal.tjfxtras.datautils;
 
 import javafx.util.StringConverter;
